[haproxy][redis] Shutdown connections after failed check
After a failover, checks will fail for the former leader node.

Without shutting down open connections after such an event, applications
can, depending on their client-library implementation, get stuck with open
connections to the former leader, which is now a read-only follower.
This in turn leads to read-only errors for write operations, as long as
the application does not initiate a re-connect.

We should not rely on client-libraries initiating re-connects for such
events when we abstract the Sentinel away with HAProxy.

Signed-off-by: Tobias Nehrlich <tobi@vshn.ch>