Add correct backend check for the metrics backend
Without the `init-addr none` option, haproxy will not start up if the
backend doesn't exist yet (i.e is not resolvable yet, because the pod
doesn't exist yet).
Furthermore setting sane hold times for the resolver ensures, that when
a backend pod is restarted, that haproxy will get the new IP address as
soon as possible.

Signed-off-by: Nicolas Bigler <nicolas.bigler@vshn.ch>